+++ packages/lib/index.ts
@@ -1,8 +1,5 @@
 export { default as isPrismaObj, isPrismaObjOrUndefined } from "./isPrismaObj";
 export * from "./isRecurringEvent";
-export * from "./isBookingLimits";
-export * from "./isDurationLimits";
 export * from "./isEventTypeColor";
-export * from "./validateIntervalLimitOrder";
 export * from "./schedules";
 export * from "./event-types";
