+++ packages/lib/server/index.ts
@@ -1,5 +1,3 @@
-export { checkBookingLimits, checkBookingLimit } from "./checkBookingLimits";
-export { checkDurationLimits, checkDurationLimit } from "./checkDurationLimits";
 export { defaultHandler } from "./defaultHandler";
 export { defaultResponder } from "./defaultResponder";
 export { getLuckyUser } from "./getLuckyUser";
